requirements.txt
wheel
flask
gunicorn
google-cloud-datastore
google-cloud-translate
rsa
requests_oauthlib
